move contributor section to control/ package
