+++ jvm/controls/src/main/kotlin/org/wycliffeassociates/otter/jvm/controls/ContributorInfo.kt
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU General Public License
  * along with Orature.  If not, see <https://www.gnu.org/licenses/>.
  */
-package org.wycliffeassociates.otter.jvm.workbookapp.ui.components
+package org.wycliffeassociates.otter.jvm.controls
 
 import javafx.beans.binding.Bindings
 import javafx.beans.property.SimpleIntegerProperty
